backend/pyproject.toml
[build-system]
requires = ["setuptools>=72.0", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "mixmatch"
dynamic = ["version"]
requires-python = ">=3.11, <3.12"
dependencies = [
    "alembic==1.16.1",
    "argon2-cffi==23.1.0",
    "cachetools==6.0.0",
    "celery[sqlalchemy]==5.5.2",
    "cron-validator==1.0.8",
    "click==8.2.1",
    "essentia==2.1b6.dev1110",
    "fastapi==0.115.12",
    "fastapi-pagination==0.13.1",
    "mutagen==1.47.0",
    "numpy<2.0.0",  # required for essentia==2.1b6.dev1110
    "prettytable==3.16.0",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.9.1",
    "python-multipart==0.0.20",
    "sqlmodel==0.0.24",
    "uvicorn[standard]==0.34.3"
]

[dependency-groups]
dev = ["setuptools", "setuptools-scm", {include-group = "test"}]
test = ["httpx==0.28.1", "pytest==8.3.5", "tox==4.26.0", "Faker==37.3.0"]

[project.scripts]
mixmatch = "mixmatch.__main__:main"

[tool.setuptools.packages.find]
include = ["mixmatch", "mixmatch.*"]

[tool.setuptools_scm]
root = ".."

[tool.tox]
requires = ["tox>=4.23"]
env_list = ["3.11"]

[tool.tox.env_run_base]
description = "run unit tests"
dependency_groups = ["test"]
set_env = {BASE_DIRECTORY="{envtmpdir}", IMAGE_DIRECTORY="{envtmpdir}/images", MUSIC_DIRECTORY="{envtmpdir}"}
commands = [["pytest", "tests"]]
